smooth walking when changing plan (#6)
when a character goes from one walking on one path to walking on another, we can make the transition smooth instead of rounding to the nearest integer location first.
we do this by using the current fractional position as the start of the new path, adding the little extra distance to every path, and the integer we're aiming for now takes into account our orientation.

this doesn't affect the NPCs very much, since they only do one walk at a time. but it helps a lot when we allow interaction to override paths. when you transition from walking on one path to another, there's no longer a little jump.