fix prompt so agents stop calling each other [Object object] (#7)
fix a chatgpt prompt in some small ways:

we're defining nearbyPlayersNames as the string "[Object object]" in most cases. sometimes the agents call each other by this name.
Screenshot 2023-08-23 at 12 44 52 PM

the timestamps are in unix epoch seconds, which probably isn't the easiest for chatgpt to parse, and definitely not as easy for a human to parse if they are printing out prompts.

if two agents haven't talked to each other before, the timestamp will be the string "undefined".

there's a trailing } curly brace in the string