make Kurt more fun to talk to (#13)
the conversations can get pretty repetitive. this is even true if a human tries to steer the conversation, because the characters fixate on their identity. So if you talk about X to Alice, she will say "ah the mysteries of X. isn't the universe fascinating". Kurt is especially prone to this because he is always alluding to his secret which he won't ever tell you but he always talks about. Let's make him into a fountain of fun facts.