refactor walking into mutation (#23)
refactor the walking and agent completion code so walkToTarget can be called from multiple different mutations. And the agent completion code can be a single mutation instead of a function called from an action, which makes it one atomic mutation with a large read set instead of two.

this will come in handy if we want to call these functions from other mutations, especially when we add human interaction.