Integrate frontend UI to aitown,select map working
I have restructured the src folder so now it's using react router. All the code related to the config page is
in config folder. The aitown page is in GamePage.tsx and APP.tsx is used to handle react router.

Currently when click the save button at the end(on the music page), only the map handler logic is working.